silx.gui.data.DataViews: Break it down in several modules
