+++ src/silx/gui/data/modes.py
@@ -1,0 +1,25 @@
+"""DataViewer modes"""
+
+EMPTY_MODE = 0
+PLOT1D_MODE = 10
+RECORD_PLOT_MODE = 15
+IMAGE_MODE = 20
+PLOT2D_MODE = 21
+COMPLEX_PLOT2D_MODE = 22
+PLOT3D_MODE = 30
+RAW_MODE = 40
+RAW_ARRAY_MODE = 41
+RAW_RECORD_MODE = 42
+RAW_SCALAR_MODE = 43
+RAW_HEXA_MODE = 44
+STACK_MODE = 50
+HDF5_MODE = 60
+NXDATA_MODE = 70
+NXDATA_INVALID_MODE = 71
+NXDATA_SCALAR_MODE = 72
+NXDATA_CURVE_MODE = 73
+NXDATA_XYVSCATTER_MODE = 74
+NXDATA_IMAGE_MODE = 75
+NXDATA_STACK_MODE = 76
+NXDATA_VOLUME_MODE = 77
+NXDATA_VOLUME_AS_STACK_MODE = 78
